fix: split extrisinc
