+++ src/utils/hooks/transaction/extrinsics/index.ts
@@ -1,0 +1,1 @@
+export { getExtrinsic, getStatus } from './extrinsics';
